IntelliJ optimizations

+++ src/me/nov/threadtear/analysis/stack/ConstantAnalyzer.java
@@ -282,17 +282,17 @@ public class ConstantAnalyzer implements Opcodes {
       Object lowerVal = low.getValue();
       switch (op) {
       case IF_ICMPEQ:
-        return ((Integer) upperVal).equals(lowerVal) ? 1 : -1;
+        return upperVal.equals(lowerVal) ? 1 : -1;
       case IF_ICMPNE:
-        return ((Integer) upperVal).equals(lowerVal) ? -1 : 1;
+        return upperVal.equals(lowerVal) ? -1 : 1;
       case IF_ICMPLT:
-        return ((Integer) lowerVal) < ((Integer) upperVal) ? 1 : -1;
+        return (Integer) lowerVal < (Integer) upperVal ? 1 : -1;
       case IF_ICMPGE:
-        return ((Integer) lowerVal) >= ((Integer) upperVal) ? 1 : -1;
+        return (Integer) lowerVal >= (Integer) upperVal ? 1 : -1;
       case IF_ICMPGT:
-        return ((Integer) lowerVal) > ((Integer) upperVal) ? 1 : -1;
+        return (Integer) lowerVal > (Integer) upperVal ? 1 : -1;
       case IF_ICMPLE:
-        return ((Integer) lowerVal) <= ((Integer) upperVal) ? 1 : -1;
+        return (Integer) lowerVal <= (Integer) upperVal ? 1 : -1;
       case IF_ACMPEQ:
         return up.equals(low) ? 1 : -1;
       case IF_ACMPNE:
